encapsulate rethnet provider in its own module
addresses review comment https://github.com/NomicFoundation/hardhat/pull/4095#discussion_r1271048119
