+++ packages/hardhat-core/test/internal/hardhat-network/helpers/useProvider.ts
@@ -1,11 +1,6 @@
-import type { Client as ClientT } from "undici";
-import { normalize, resolve as resolvePath } from "node:path";
-import { spawn } from "node:child_process";
-
 import { HardhatNetworkChainsConfig } from "../../../../src/types/config";
 import { defaultHardhatNetworkParams } from "../../../../src/internal/core/config/default-config";
 import { BackwardsCompatibilityProviderAdapter } from "../../../../src/internal/core/providers/backwards-compatibility";
-import { HttpProvider } from "../../../../src/internal/core/providers/http";
 import { JsonRpcServer } from "../../../../src/internal/hardhat-network/jsonrpc/server";
 import {
   ForkConfig,
@@ -31,6 +26,7 @@ import {
   DEFAULT_USE_JSON_RPC,
 } from "./providers";
 import { sleep } from "./sleep";
+import { spawnRethnetProvider } from "./spawnRethnetProvider";
 
 declare module "mocha" {
   interface Context {
@@ -125,74 +121,17 @@ export function useProvider({
     }
 
     if (rethnetBinary !== undefined) {
-      this.rethnetProcess = spawn(normalize(rethnetBinary), ["node", "-vv"]);
+      const { childProcess, processExit, httpProvider } =
+        spawnRethnetProvider(rethnetBinary);
 
-      let stdout = "";
-      this.rethnetProcess.stdout.on("data", (data: any) => {
-        stdout += data.toString();
-        console.log(`rethnet subprocess ${this.rethnetProcess.pid}: ${data}`);
-      });
-
-      let stderr = "";
-      this.rethnetProcess.stderr.on("data", (data: any) => {
-        stderr += data.toString();
-        console.log(`rethnet subprocess ${this.rethnetProcess.pid}: ${data}`);
-      });
-
-      function outputForError() {
-        return `stdout:\n${stdout}\nstderr:\n${stderr}`;
-      }
-
-      const exit = new Promise<void>((resolve, reject) => {
-        this.rethnetProcess.on("exit", (code: number, signal: string) => {
-          if (signal === "SIGKILL") {
-            console.log("kill");
-            resolve();
-          } else {
-            reject(
-              new Error(
-                `rethnet process closed unexpectedly. return code: ${code}. signal: ${signal}. ${outputForError()}`
-              )
-            );
-          }
-        });
-      });
-
-      const error = new Promise((_resolve, reject) => {
-        this.rethnetProcess.on("error", (err: Error) => {
-          if (err.message.includes("ENOENT")) {
-            reject(
-              new Error(
-                `Rethnet executable not found at ${resolvePath(rethnetBinary)}`
-              )
-            );
-          } else {
-            reject(
-              new Error(`Rethnet subprocess error: ${err}. ${outputForError()}`)
-            );
-          }
-        });
-      });
+      this.rethnetProcess = childProcess;
 
       // wait for the server to initialize:
       await sleep(250);
 
-      const { Client } = require("undici") as { Client: typeof ClientT };
-      const url = "http://127.0.0.1:8545";
-      this.provider = new BackwardsCompatibilityProviderAdapter(
-        new HttpProvider(
-          url,
-          "rethnet",
-          {},
-          20000,
-          new Client(url, {
-            keepAliveTimeout: 10,
-            keepAliveMaxTimeout: 10,
-          })
-        )
-      );
+      this.provider = new BackwardsCompatibilityProviderAdapter(httpProvider);
 
-      await Promise.race([sleep(2000), exit, error]);
+      await processExit;
     }
   });
